requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
psycopg2-binary==2.9.9
pandas==2.0.3
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
prometheus-client==0.17.1
